--- src/containers/productList.js
@@ -1,2 +1,0 @@
-import React from 'react';
-import PropTypes from 'prop-types';
